python-worker/requirements.txt
sentry-sdk>=2.46.0
requests>=2.31.0

python-dotenv>=1.0.0
